requirements.txt
blinker~=1.6.2
celery~=5.3.1
datetime~=5.0
Flask~=2.2
Flask-Mail~=0.9.1
python-dotenv~=1.0.0
redis~=4.5.4
requests~=2.28
requests-cache~=0.9
skyfield~=1.45
numba~=0.56
numpy~=1.21.0
urllib3~=1.26.14
